requirements.txt
python-osc
DMXEnttecPro
pypubsub
pyyaml
singleton-decorator